13course_digest/requirements.txt
faster-whisper>=1.0.0    # 本地语音转文字（OpenAI Whisper 优化版，完全免费）
python-pptx>=0.6.23      # PPTX 文件解析
pdfplumber>=0.10.0       # PDF 文件解析
groq>=0.11.0             # Groq API 客户端（免费额度，Llama-3.3-70B）
dashscope>=1.20.0         # 阿里巴巴百炼 API 客户端（千问）
PyYAML>=6.0              # YAML 配置文件解析
tqdm>=4.66.0             # 命令行进度条
